Restructure deployment package and modules
