+++ boss/api/deployment/deployer.py
@@ -14,11 +14,11 @@ def import_preset(config):
         )
         # TODO: In future release, don't import deployment tasks
         # unless the preset is set. (BC Break)
-        from boss.api.deployment import remote_source as module
+        from .preset import remote_source as module
     elif preset == constants.PRESET_REMOTE_SOURCE:
-        from boss.api.deployment import remote_source as module
+        from .preset import remote_source as module
     elif preset == constants.PRESET_FRONTEND:
-        from boss.api.deployment import frontend as module
+        from .preset import frontend as module
     else:
         halt('Unsupported boss preset "{}".'.format(preset))
